requirements.txt
fastapi
uvicorn
pdfplumber
python-docx
langchain
langchain-community
serpapi
python-dotenv
